Battle Tank - Lecture 172 - Fixing 3rd Person Camera Rotation
We added a Scene object that we used as a "Gimbal" support (google it) which detaches the spring arm from the tank and fixes that rotation issue. We also changed a Spring arm property to avoid inherit the "Roll" value from the tank, to keep the horizon while going through the hills. The gimbal provides the movement for the azimuth, and the spring arms provides the elevation